Use registry and autodiscovery for exports
