+++ meinberlin/apps/documents/views.py
@@ -1,13 +1,11 @@
 from django.http import Http404
 from django.http.response import HttpResponseRedirect
-from django.utils.html import strip_tags
 from django.utils.translation import ugettext_lazy as _
 from django.views import generic
 
 from adhocracy4.rules import mixins as rules_mixins
 from meinberlin.apps.contrib.views import ProjectContextDispatcher
 from meinberlin.apps.dashboard2 import mixins as dashboard_mixins
-from meinberlin.apps.exports import views as export_views
 
 from . import models
 
@@ -46,34 +44,7 @@ class ChapterDetailView(ProjectContextDispatcher,
         return models.Chapter.objects.filter(module=self.module)
 
 
-class DocumentExportView(export_views.AbstractXlsxExportView,
-                         export_views.ItemExportWithCommentsMixin):
-
-    PARAGRAPH_TEXT_LIMIT = 100
-
-    def get_header(self):
-        return map(str, [_('Chapter'), _('Paragraph'), _('Comments')])
-
-    def export_rows(self):
-        chapters = models.Chapter.objects.filter(module=self.module)
-
-        for chapter in chapters:
-            yield [chapter.name, '', self.get_comments_data(chapter)]
-            for paragraph in chapter.paragraphs.all():
-                yield [chapter.name,
-                       self.get_paragraph_data(paragraph),
-                       self.get_comments_data(paragraph)]
-
-    def get_paragraph_data(self, paragraph):
-        if paragraph.name:
-            return paragraph.name
-
-        text = strip_tags(paragraph.text).strip()
-        return text[0:self.PARAGRAPH_TEXT_LIMIT] + ' ...'
-
-
 class DocumentDetailView(ChapterDetailView):
-    exports = [(_('Documents with comments'), DocumentExportView)]
 
     def get_object(self):
         first_chapter = models.Chapter.objects \
