+++ meinberlin/apps/ideas/views.py
@@ -10,7 +10,6 @@ from adhocracy4.filters.filters import FreeTextFilter
 from adhocracy4.rules import mixins as rules_mixins
 from meinberlin.apps.contrib import filters
 from meinberlin.apps.contrib.views import ProjectContextDispatcher
-from meinberlin.apps.exports import views as export_views
 
 from . import forms
 from . import models
@@ -51,25 +50,9 @@ class AbstractIdeaListView(ProjectContextDispatcher,
     paginate_by = 15
 
 
-class IdeaExportView(export_views.ItemExportView,
-                     export_views.ItemExportWithRatesMixin,
-                     export_views.ItemExportWithCommentCountMixin,
-                     export_views.ItemExportWithCommentsMixin):
-    model = models.Idea
-    fields = ['name', 'description', 'creator', 'created']
-
-    def get_queryset(self):
-        return super().get_queryset() \
-            .filter(module=self.module)\
-            .annotate_comment_count()\
-            .annotate_positive_rating_count()\
-            .annotate_negative_rating_count()
-
-
 class IdeaListView(AbstractIdeaListView):
     model = models.Idea
     filter_set = IdeaFilterSet
-    exports = [(_('Ideas with comments'), IdeaExportView)]
 
     def get_queryset(self):
         return super().get_queryset()\
