+++ meinberlin/apps/mapideas/views.py
@@ -2,7 +2,6 @@ from django.utils.translation import ugettext_lazy as _
 
 from adhocracy4.filters import filters as a4_filters
 from meinberlin.apps.contrib import filters
-from meinberlin.apps.exports import views as export_views
 from meinberlin.apps.ideas import views as idea_views
 
 from . import forms
@@ -31,26 +30,9 @@ class MapIdeaFilterSet(a4_filters.DefaultsFilterSet):
         fields = ['category']
 
 
-class MapIdeaExportView(export_views.ItemExportView,
-                        export_views.ItemExportWithRatesMixin,
-                        export_views.ItemExportWithCommentCountMixin,
-                        export_views.ItemExportWithCommentsMixin,
-                        export_views.ItemExportWithLocationMixin):
-    model = models.MapIdea
-    fields = ['name', 'description', 'creator', 'created']
-
-    def get_queryset(self):
-        return super().get_queryset() \
-            .filter(module=self.module)\
-            .annotate_comment_count()\
-            .annotate_positive_rating_count()\
-            .annotate_negative_rating_count()
-
-
 class MapIdeaListView(idea_views.AbstractIdeaListView):
     model = models.MapIdea
     filter_set = MapIdeaFilterSet
-    exports = [(_('Ideas with location and comments'), MapIdeaExportView)]
 
     def dispatch(self, request, **kwargs):
         self.mode = request.GET.get('mode', 'map')
