+++ meinberlin/apps/topicprio/views.py
@@ -7,7 +7,6 @@ from adhocracy4.filters import widgets as filters_widgets
 from adhocracy4.filters.filters import FreeTextFilter
 from meinberlin.apps.contrib import filters
 from meinberlin.apps.dashboard2 import mixins
-from meinberlin.apps.exports import views as export_views
 from meinberlin.apps.ideas import views as idea_views
 
 from . import forms
@@ -40,25 +39,9 @@ class TopicFilterSet(a4_filters.DefaultsFilterSet):
         fields = ['search', 'category']
 
 
-class TopicExportView(export_views.ItemExportView,
-                      export_views.ItemExportWithRatesMixin,
-                      export_views.ItemExportWithCommentCountMixin,
-                      export_views.ItemExportWithCommentsMixin):
-    model = models.Topic
-    fields = ['name', 'description', 'creator', 'created']
-
-    def get_queryset(self):
-        return super().get_queryset() \
-            .filter(module=self.module)\
-            .annotate_comment_count()\
-            .annotate_positive_rating_count()\
-            .annotate_negative_rating_count()
-
-
 class TopicListView(idea_views.AbstractIdeaListView):
     model = models.Topic
     filter_set = TopicFilterSet
-    exports = [(_('Topics with comments'), TopicExportView)]
 
     def get_queryset(self):
         return super().get_queryset()\
